requirements.txt
# Core libraries
torch
transformers
datasets
faiss-cpu
spacy
jinja2
sentence-transformers
accelerate
bitsandbytes
chromadb
# Utilities & evaluation
scikit-learn
evaluate
pandas
tqdm
rich
numpy

# LLM APIs & demo
openai
streamlit
flask
nltk
matplotlib
json

# Configuration
python-dotenv
